tests: move user creation function/class at the root level
Move:
- create_user
- WisdomServiceAPITestCaseBase
- WisdomServiceAPITestCaseBaseOIDC

At the root level in `ansible_ai_connect.test_utils` to simplify the reusablity.
